+++ src/app/app.module.ts
@@ -28,7 +28,7 @@ import { MonthComponent } from './month/month.component';
 import { WeekComponent } from './week/week.component';
 import { CalendarContainerComponent } from './calendar-container/calendar-container.component';
 import { SearchBarComponent } from './search-bar/search-bar.component';
-import { ModalComponent } from './sidebar-popup/sidebar-popup.component';
+import { ModalComponent } from './modal/modal.component';
 
 
 const appRoutes: Routes = [
